+++ src/tasks.ts
@@ -138,10 +138,14 @@ const deployUNSTask = {
   run: async (ctx: Deployer, dependencies: DependenciesMap) => {
     const { owner } = ctx.accounts;
     const [
+      CNSRegistry,
+      RootChainManager,
       MintingController,
       URIPrefixController,
       Resolver,
     ] = unwrapDependencies(dependencies, [
+      ArtifactName.CNSRegistry,
+      ArtifactName.RootChainManager,
       ArtifactName.MintingController,
       ArtifactName.URIPrefixController,
       ArtifactName.Resolver,
@@ -184,7 +188,12 @@ const deployUNSTask = {
       await verify(ctx, mintingManager.address, []);
     }
 
-    const registryInitTx = await unsRegistry.connect(owner).initialize(mintingManager.address);
+    const registryInitTx = await unsRegistry.connect(owner).initialize(
+      mintingManager.address,
+      CNSRegistry.address,
+      RootChainManager.address,
+      ZERO_ADDRESS,
+    );
     await registryInitTx.wait();
 
     const forwarder = await ctx.artifacts.MintingManagerForwarder
@@ -308,84 +317,6 @@ const deployUNSProxyReaderTask: Task = {
   },
 };
 
-const configureCNSTask = {
-  tags: ['uns_config_cns', 'full'],
-  priority: 20,
-  run: async (ctx: Deployer, dependencies: DependenciesMap) => {
-    const { owner } = ctx.accounts;
-
-    const [
-      CNSRegistry,
-      MintingController,
-      URIPrefixController,
-      UNSRegistry,
-      MintingManager,
-    ] = unwrapDependencies(dependencies, [
-      ArtifactName.CNSRegistry,
-      ArtifactName.MintingController,
-      ArtifactName.URIPrefixController,
-      ArtifactName.UNSRegistry,
-      ArtifactName.MintingManager,
-    ]);
-
-    const mintingController = ctx.artifacts.MintingController
-      .attach(MintingController.address)
-      .connect(owner);
-
-    if (!(await mintingController.isMinter(MintingManager.address))) {
-      await mintingController.addMinter(MintingManager.address);
-    }
-
-    const uriPrefixController = ctx.artifacts.URIPrefixController
-      .attach(URIPrefixController.address)
-      .connect(owner);
-
-    if (!(await uriPrefixController.isWhitelisted(MintingManager.address))) {
-      await uriPrefixController.addWhitelisted(MintingManager.address);
-    }
-
-    const unsRegistry = ctx.artifacts.UNSRegistry
-      .attach(UNSRegistry.address)
-      .connect(owner);
-
-    await unsRegistry.setCNSRegistry(CNSRegistry.address);
-
-    // Set tokenURI prefix only for Sandbox
-    if (network.config.chainId === 1337) {
-      const mintingManager = ctx.artifacts.MintingManager
-        .attach(MintingManager.address)
-        .connect(owner);
-      await mintingManager.setTokenURIPrefix('https://example.com/');
-    }
-  },
-  ensureDependencies: (ctx: Deployer, config?: UnsNetworkConfig): DependenciesMap => {
-    config = merge(ctx.getDeployConfig(), config);
-
-    const {
-      CNSRegistry,
-      MintingController,
-      URIPrefixController,
-      UNSRegistry,
-      MintingManager,
-    } = config.contracts || {};
-    const dependencies = {
-      CNSRegistry,
-      MintingController,
-      URIPrefixController,
-      UNSRegistry,
-      MintingManager,
-    };
-
-    for (const [key, value] of Object.entries(dependencies)) {
-      if (!value || !value.address) {
-        throw new Error(`${key} contract not found for network ${network.config.chainId}`);
-      }
-    }
-
-    return dependencies;
-  },
-};
-
 const deployMMForwarderTask = {
   tags: ['uns_mm_forwarder'],
   priority: 100,
@@ -458,6 +389,7 @@ const upgradeUNSRegistryTask: Task = {
     return dependencies;
   },
 };
+
 
 const upgradeMintingManagerTask: Task = {
   tags: ['upgrade_minting_manager'],
@@ -548,52 +480,6 @@ const upgradeProxyReaderTask = {
 };
 
 /**
- * The task configure CNS domain migration to UNS.
- * It is required for post-upgrade configuration.
- */
-const configureCnsMigrationTask: Task = {
-  tags: ['uns_config_cns_migration'],
-  priority: 100,
-  run: async (ctx: Deployer, dependencies: DependenciesMap) => {
-    const [ CNSRegistry, UNSRegistry ] = unwrapDependencies(
-      dependencies,
-      [
-        ArtifactName.CNSRegistry,
-        ArtifactName.UNSRegistry,
-      ],
-    );
-
-    const { owner } = ctx.accounts;
-
-    const unsRegistry = ctx.artifacts.UNSRegistry
-      .attach(UNSRegistry.address)
-      .connect(owner);
-
-    await unsRegistry.setCNSRegistry(CNSRegistry.address);
-  },
-  ensureDependencies: (ctx: Deployer, config?: UnsNetworkConfig) => {
-    config = merge(ctx.getDeployConfig(), config);
-
-    const {
-      CNSRegistry,
-      UNSRegistry,
-    } = config.contracts || {};
-    const dependencies = {
-      CNSRegistry,
-      UNSRegistry,
-    };
-
-    for (const [key, value] of Object.entries(dependencies)) {
-      if (!value || !value.address) {
-        throw new Error(`${key} contract not found for network ${network.config.chainId}`);
-      }
-    }
-
-    return dependencies;
-  },
-};
-
-/**
  * The task deploys Polygon POS Bridge contracts.
  * It is required for emulation of Bridge for test environments.
  */
@@ -645,46 +531,6 @@ const deployPolygonPosBridgeTask: Task = {
   },
 };
 
-const configurePolygonPosBridgeTask: Task = {
-  tags: ['uns_config_polygon_pos_bridge'],
-  priority: 140,
-  run: async (ctx: Deployer, dependencies: DependenciesMap) => {
-    const { owner } = ctx.accounts;
-
-    const [ UNSRegistry, RootChainManager ] = unwrapDependencies(dependencies,
-      [
-        ArtifactName.UNSRegistry,
-        ArtifactName.RootChainManager,
-      ],
-    );
-
-    const unsRegistry = ctx.artifacts.UNSRegistry
-      .attach(UNSRegistry.address)
-      .connect(owner);
-
-    await unsRegistry.setRootChainManager(RootChainManager.address);
-  },
-  ensureDependencies: (ctx: Deployer, config?: UnsNetworkConfig) => {
-    config = merge(ctx.getDeployConfig(), config);
-
-    const {
-      UNSRegistry,
-      RootChainManager,
-    } = config.contracts || {};
-    const dependencies = {
-      UNSRegistry,
-      RootChainManager,
-    };
-
-    for (const [key, value] of Object.entries(dependencies)) {
-      if (!value || !value.address) {
-        throw new Error(`${key} contract not found for network ${network.config.chainId}`);
-      }
-    }
-
-    return dependencies;
-  },
-};
 
 const configureDotCoinTask: Task = {
   tags: ['temp_configure_dotcoin'],
@@ -731,152 +577,15 @@ const configureDotCoinTask: Task = {
   },
 };
 
-const configureReconfigureTldL1Task = {
-  tags: ['temp_reconfigure_tld_l1'],
-  priority: 115,
-  run: async (ctx: Deployer, dependencies: DependenciesMap) => {
-    const { owner } = ctx.accounts;
-    const MintingManager = unwrap(dependencies, ArtifactName.MintingManager);
-
-    const mintingManager = ctx.artifacts.MintingManager.attach(
-      MintingManager.address,
-    ).connect(owner);
-
-    // Burn TLD tokens
-    const tldTokens = [
-      '0x7674e7282552c15f203b9c4a6025aeaf28176ef7f5451b280f9bada3f8bc98e2', // .coin
-      '0x0f4a10a4f46c288cea365fcf45cccf0e9d901b945b9829ccdb54c10dc3cb7a6f', // .crypto
-      '0x1e3f482b3363eb4710dae2cb2183128e272eafbe137f686851c1caea32502230', // .wallet
-      '0x241e7e2b7fd7333b3c0c049b326316b811af0c01cfc0c7a90b466fda3a70fc2d', // .x
-      '0xb75cf4f3d8bc3deb317ed5216d898899d5cc6a783f65f6768eb9bcb89428670d', // .nft
-      '0x4118ebbd893ecbb9f5d7a817c7d8039c1bd991b56ea243e2ae84d0a1b2c950a7', // .blockchain
-      '0x042fb01c1e43fb4a32f85b41c821e17d2faeac58cfc5fb23f80bc00c940f85e3', // .bitcoin
-      '0x5c828ec285c0bf152a30a325b3963661a80cb87641d60920344caf04d4a0f31e', // .888
-      '0xb5f2bbf81da581299d4ff7af60560c0ac854196f5227328d2d0c2bb0df33e553', // .dao
-      '0xd81bbfcee722494b885e891546eeac23d0eedcd44038d7a2f6ef9ec2f9e0d239', // .zil
-      '0xed9ce6b49a0e2c56c57c86795b131bd6df792312183994c3cf3de1516cfe92d6', // .polygon
-      '0x92bba949890cd44a226a8ce54135cf86538cd6c5ca0ccf41877102fd718cc8aa', // .unstoppable
-    ];
-    await mintingManager.burnTLDL1(tldTokens);
-  },
-  ensureDependencies: (ctx: Deployer, config?: UnsNetworkConfig) => {
-    config = merge(ctx.getDeployConfig(), config);
-
-    const { chainId } = unwrap(network.config, 'chainId');
-
-    if (![1, 5].includes(chainId)) {
-      throw new Error('Current network configuration does not support burning TLD');
-    }
-
-    const { MintingManager } = config.contracts || {};
-    const dependencies = { MintingManager };
-    for (const [key, value] of Object.entries(dependencies)) {
-      if (!value || !value.address) {
-        throw new Error(`${key} contract not found for network ${chainId}`);
-      }
-    }
-
-    return dependencies;
-  },
-};
-
-const configureReconfigureTldL2Task: Task = {
-  tags: ['temp_reconfigure_tld_l2'],
-  priority: 120,
-  run: async (ctx: Deployer, dependencies: DependenciesMap) => {
-    const { owner } = ctx.accounts;
-    const MintingManager = unwrap(dependencies, ArtifactName.MintingManager);
-
-    const mintingManager = ctx.artifacts.MintingManager.attach(
-      MintingManager.address,
-    ).connect(owner);
-
-    // Transfer TLD tokens ownership to MintingManager
-    const tldTokens = [
-      '0x0f4a10a4f46c288cea365fcf45cccf0e9d901b945b9829ccdb54c10dc3cb7a6f', // .crypto
-      '0x1e3f482b3363eb4710dae2cb2183128e272eafbe137f686851c1caea32502230', // .wallet
-      '0x241e7e2b7fd7333b3c0c049b326316b811af0c01cfc0c7a90b466fda3a70fc2d', // .x
-      '0xb75cf4f3d8bc3deb317ed5216d898899d5cc6a783f65f6768eb9bcb89428670d', // .nft
-      '0x4118ebbd893ecbb9f5d7a817c7d8039c1bd991b56ea243e2ae84d0a1b2c950a7', // .blockchain
-      '0x042fb01c1e43fb4a32f85b41c821e17d2faeac58cfc5fb23f80bc00c940f85e3', // .bitcoin
-      '0x5c828ec285c0bf152a30a325b3963661a80cb87641d60920344caf04d4a0f31e', // .888
-      '0xb5f2bbf81da581299d4ff7af60560c0ac854196f5227328d2d0c2bb0df33e553', // .dao
-      '0xd81bbfcee722494b885e891546eeac23d0eedcd44038d7a2f6ef9ec2f9e0d239', // .zil
-      '0xed9ce6b49a0e2c56c57c86795b131bd6df792312183994c3cf3de1516cfe92d6', // .polygon
-      '0x92bba949890cd44a226a8ce54135cf86538cd6c5ca0ccf41877102fd718cc8aa', // .unstoppable
-      '0xa18784bb78ee0f577251fb21ad5cac7a140ab47e9414e3c7af5125e3e1d28923',  // .klever
-    ];
-    await mintingManager.moveTLDOwnershipL2(tldTokens);
-  },
-  ensureDependencies: (ctx: Deployer, config?: UnsNetworkConfig) => {
-    config = merge(ctx.getDeployConfig(), config);
-
-    const { chainId } = unwrap(network.config, 'chainId');
-    if (![137, 80001].includes(chainId)) {
-      throw new Error('Current network configuration does not support moving TLD ownership');
-    }
-
-    const { MintingManager } = config.contracts || {};
-    const dependencies = { MintingManager };
-    for (const [key, value] of Object.entries(dependencies)) {
-      if (!value || !value.address) {
-        throw new Error(`${key} contract not found for network ${chainId}`);
-      }
-    }
-
-    return dependencies;
-  },
-};
-
-const deployDotCoinBurnerTask = {
-  tags: ['dot_coin_burner', 'full'],
-  priority: 200,
-  run: async (ctx: Deployer, dependencies: DependenciesMap) => {
-    const { owner } = ctx.accounts;
-    const UNSRegistry = unwrap(dependencies, ArtifactName.UNSRegistry);
-
-    const dotCoinBurner = await ctx.artifacts.DotCoinBurner
-      .connect(owner)
-      .deploy(UNSRegistry.address);
-    await ctx.saveContractConfig(UnsContractName.DotCoinBurner, dotCoinBurner);
-    await dotCoinBurner.deployTransaction.wait();
-    await verify(ctx, dotCoinBurner.address, [UNSRegistry.address]);
-  },
-  ensureDependencies: (ctx: Deployer, config?: UnsNetworkConfig) => {
-    config = merge(ctx.getDeployConfig(), config);
-
-    const {
-      UNSRegistry,
-    } = config.contracts || {};
-    const dependencies = {
-      UNSRegistry,
-    };
-
-    for (const [key, value] of Object.entries(dependencies)) {
-      if (!value || !value.address) {
-        throw new Error(`${key} contract not found for network ${network.config.chainId}`);
-      }
-    }
-
-    return dependencies;
-  },
-};
-
 export const tasks: Task[] = [
   deployCNSTask,
   deployCNSForwardersTask,
   deployUNSTask,
   deployUNSProxyReaderTask,
-  configureCNSTask,
   deployMMForwarderTask,
   upgradeUNSRegistryTask,
   upgradeMintingManagerTask,
   upgradeProxyReaderTask,
-  configureCnsMigrationTask,
   deployPolygonPosBridgeTask,
-  configurePolygonPosBridgeTask,
   configureDotCoinTask,
-  configureReconfigureTldL1Task,
-  configureReconfigureTldL2Task,
-  deployDotCoinBurnerTask,
 ];