+++ test/ProxyReader.test.ts
@@ -35,7 +35,7 @@ describe('ProxyReader', () => {
 
     // deploy UNS
     unsRegistry = await new UNSRegistry__factory(coinbase).deploy();
-    await unsRegistry.initialize(coinbase.address);
+    await unsRegistry.initialize(coinbase.address, ZERO_ADDRESS, ZERO_ADDRESS, ZERO_ADDRESS);
     await unsRegistry.setTokenURIPrefix('/');
 
     // deploy CNS